update form regis pegawai,session, dan hak akses

+++ delete-departemen.php
@@ -1,12 +1,18 @@
 <?php
 
 require 'connection.php';
+//Proteksi: Jika belum login dan bukan admin di tendang ke halaman login
+if (!isset($_SESSION['login']) || $_SESSION['level'] !== 'admin') {
+    header("Location: login.php");
+    exit;
+}
+
 $id = $_GET["id"];
 
 if(isset($id)){
     try{
         mysqli_query($connection,"DELETE FROM departemen WHERE id = '$id'");
-        header("Location:list-departemen.php");
+        echo "<script>alert('Data Berhasil Dihapus!'); window.location='list-departemen.php';</script>";
     }catch(Exception $e){
         echo "Failed delete data " .$e -> getMessage();
     }
